+++ CHANGELOG.md
@@ -1,8 +1,16 @@
-## 1.0.1
+## 1.0.0+4
+- Some typos in LorDeckEncoder.dart fixed
+  
+## 1.0.0+3
+- Some typos in LorDeckEncoder.dart fixed
+  
+## 1.0.0+2
+- Some typos in LorDeckEncoder.dart fixed
+  
+## 1.0.0+1
 - missing ; added
 - example folder and example file added for show how to use code
   
 ## 1.0.0
-
 - Package upload.
 - Required Classes, functions created and Tested by UnitTest.